Logging out contents of files
Logging out the contents of each file. Currently the contents are logging correctly, but the file name seems to be stuck at 03. Just realised that this is because the readFile() is happing asynchronously, so when the result is returned the loop has completed and files[f] now refers to three. Will fix on next commit