Storing data in Array - Synchronous
Following on from my earlier commit, i decided to continue with the concept of storing the data from the files in a global array. Currently, the only benefit this gives me is that I don't actually log out any data from the files until they have all been read. In reality this isn't a huge benefit to me, as I am processing the .readFile() function in a way which is almost synchronous (only one readFile() function is ever qued at a time) so the files will always come back in the right order. Just completed this to finish my previous train of thought