Adding ability to ignore files
When I created the app-b and ran it I realised that the contents of app-a.js, as well as .atom-ternjs, were now printing out the the console. To fix this using my current method (i.e of checking whether the filename was equal to app.a.js) I would have had to extend the if statement significantly. Instead, I have created an ignoreFiles array, which contains the strings of the filetypes I would like to ignore. I then passed a checkIgnore() function into the if statement, and used it to return true or false, based on whether part of the filename of the file that is currently being checked contains any of the strings I am choosing to ignore i.e. .js and .atom-ternjs