Creating app-d
Trying to make a copy of app-c work but using fully asynchronous
fs.readFile() methods within a loop. If I could find a way to pass in
the name of the file to the callback, or identify which file's data I
was reading within the callback, then I'd have it sorted. Currenty
though, this app is just saving the contents of each of the files into a
new file, but in the wrong order, and all with the same filename i.e.
the last file the app came across during it's search